refactor: removed redux
