--- src/features/game/GameProvider.tsx
@@ -1,22 +1,0 @@
-"use client";
-
-import { persistor, store } from "@/store";
-import { ReactNode } from "react";
-import { Provider } from "react-redux";
-import { PersistGate } from "redux-persist/integration/react";
-
-interface Props {
-  children: ReactNode;
-}
-
-const GameProvider = ({ children }: Props) => {
-  return (
-    <Provider store={store}>
-      <PersistGate loading={null} persistor={persistor}>
-        {children}
-      </PersistGate>
-    </Provider>
-  );
-};
-
-export default GameProvider;
