+++ src/features/game/components/gamenavigator/GameNavigator.tsx
@@ -1,7 +1,7 @@
 import CountdownNavigator from "./countdown/CountdownNavigator";
 import GuessingNavigator from "./guessing/GuessingNavigator";
 import { AnimatePresence } from "framer-motion";
-import { useGameState } from "../../zstore/gameSlice";
+import { useGameState } from "../../store/gameSlice";
 
 const GameNavigator = () => {
   const status = useGameState((state) => state.round.status);
