+++ src/features/game/components/gamenavigator/guessing/GuessedState.tsx
@@ -1,6 +1,6 @@
 import { TbConfetti } from "react-icons/tb";
 import { motion } from "framer-motion";
-import { useGameState } from "@/features/game/zstore/gameSlice";
+import { useGameState } from "@/features/game/store/gameSlice";
 
 const GuessedState = () => {
   const totalPoints = useGameState((state) => state.totalPoints);
