+++ src/features/game/components/stages/InGame.tsx
@@ -2,7 +2,7 @@ import Guessing from "./Guessing";
 import { AnimatePresence } from "framer-motion";
 import CountdownNavigator from "../gamenavigator/countdown/CountdownNavigator";
 import GuessingNavigator from "../gamenavigator/guessing/GuessingNavigator";
-import { useGameState } from "../../zstore/gameSlice";
+import { useGameState } from "../../store/gameSlice";
 
 const InGame = () => {
   const { status, song } = useGameState((state) => state.round);
