--- src/hooks/index.ts
@@ -1,5 +1,0 @@
-import { AppDispatch, RootState } from "@/store";
-import { TypedUseSelectorHook, useDispatch, useSelector } from "react-redux";
-
-export const useAppDispatch: () => AppDispatch = useDispatch;
-export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
